+++ transport/http/outbound.go
@@ -169,8 +169,8 @@ type Outbound struct {
 	bothResponseError bool
 }
 
-// Name is the transport name that will be set on `transport.Request` struct.
-func (o *Outbound) Name() string {
+// TransportName is the transport name that will be set on `transport.Request` struct.
+func (o *Outbound) TransportName() string {
 	return transportName
 }