+++ transport/tchannel/outbound.go
@@ -68,8 +68,8 @@ func (t *Transport) NewSingleOutbound(addr string) *Outbound {
 	return t.NewOutbound(chooser)
 }
 
-// Name is the transport name that will be set on `transport.Request` struct.
-func (o *Outbound) Name() string {
+// TransportName is the transport name that will be set on `transport.Request` struct.
+func (o *Outbound) TransportName() string {
 	return transportName
 }